feat: centralize application configuration
