+++ services/api/tests/test_redis_connection.py
@@ -2,7 +2,7 @@ import types
 import pytest
 
 import sidetrack.api.main as main
-from sidetrack.common.config import Settings
+from sidetrack.config import Settings
 
 pytestmark = pytest.mark.unit
