+++ sidetrack/api/config.py
@@ -1,31 +1,13 @@
 """Configuration for the API service.
 
-This module wraps :mod:`sidetrack.common.config` so callers can simply import
-``sidetrack.api.config``.  Defining a dedicated module also mirrors the
-structure used by other services such as the worker and job runner.
+Historically settings lived in ``sidetrack.common.config``.  They are now
+centralised in :mod:`sidetrack.config`.  This module simply re-exports the API
+specific settings helpers for convenience.
 """
 
-from __future__ import annotations
+from sidetrack.config import ApiSettings, get_api_settings as get_settings
 
-from functools import lru_cache
-
-from sidetrack.common.config import Settings as AppSettings
-
-
-class ApiSettings(AppSettings):
-    """Settings for the API service."""
-
-    # API specific options could be added here in the future
-    pass
-
-
-@lru_cache
-def get_settings() -> ApiSettings:
-    """Return cached ``ApiSettings`` instance."""
-
-    return ApiSettings()
-
-
-# Re-export for backward compatibility with modules still importing these
-# names from ``sidetrack.common.config``.
 Settings = ApiSettings
+
+__all__ = ["ApiSettings", "Settings", "get_settings"]
+
